+++ mom6_bathy/aux.py
@@ -3,7 +3,7 @@
 from pathlib import Path
 import xarray as xr
 import numpy as np
-
+import scipy
 
 def get_mesh_dimensions(mesh):
     """Given an ESMF mesh where the grid metrics are stored in 1D (flattened) arrays,
@@ -174,16 +174,84 @@ def cell_area_rad(xv_coords, yv_coords):
 
     area = _great_circle_area(np.stack([x, y, z], axis=-1) )
     return area
-    
 
-def check_lon_range(lons, name="lon"):
-    lons = np.asarray(lons)
-    if np.all((lons >= -180) & (lons <= 180)):
-        return "[-180, 180]"
-    elif np.all((lons >= 0) & (lons <= 360)):
-        return "[0, 360]"
+
+def fill_missing_data(idata, mask, maxiter=0, stabilizer=1.0e-14, tripole=False):
+    """
+    Returns data with masked values "objectively interpolated" except where values exist or is over land. Does not work for periodic grids.
+
+    Arguments:
+    data - numpy array with nan values where there is missing data or land.
+    mask - np.array of 0 or 1, 0 for land, 1 for ocean.
+
+    Returns an array.
+    """
+
+    nj, ni = idata.shape
+    fdata = np.nan_to_num(idata, nan=0.0)
+    missing_j, missing_i = np.where(np.isnan(idata) & (mask > 0))
+
+    n_missing = missing_i.size
+
+    # ind contains column of matrix/row of vector corresponding to point [j,i]
+    ind = np.zeros(fdata.shape, dtype=int) - int(1e6)
+    ind[missing_j, missing_i] = np.arange(n_missing)
+    A = scipy.sparse.lil_matrix((n_missing, n_missing))
+    b = np.zeros((n_missing))
+    ld = np.zeros((n_missing))
+    A[range(n_missing), range(n_missing)] = 0.0
+    for n in range(n_missing):
+        j, i = missing_j[n], missing_i[n]
+        im1 = max(i - 1, 0)
+        ip1 = min(i + 1, ni - 1)
+        jm1 = max(j - 1, 0)
+        jp1 = min(j + 1, nj - 1)
+        if j > 0 and mask[jm1, i] > 0:
+            ld[n] -= 1.0
+            ij = ind[jm1, i]
+            if ij >= 0:
+                A[n, ij] = 1.0
+            else:
+                b[n] -= fdata[jm1, i]
+        if i > 0 and mask[j, im1] > 0:
+            ld[n] -= 1.0
+            ij = ind[j, im1]
+            if ij >= 0:
+                A[n, ij] = 1.0
+            else:
+                b[n] -= fdata[j, im1]
+        if i < ni - 1 and mask[j, ip1] > 0:
+            ld[n] -= 1.0
+            ij = ind[j, ip1]
+            if ij >= 0:
+                A[n, ij] = 1.0
+            else:
+                b[n] -= fdata[j, ip1]
+        if j < nj - 1 and mask[jp1, i] > 0:
+            ld[n] -= 1.0
+            ij = ind[jp1, i]
+            if ij >= 0:
+                A[n, ij] = 1.0
+            else:
+                b[n] -= fdata[jp1, i]
+        if j == nj - 1 and mask[j, ni - 1 - i] > 0 and tripole:  # Tri-polar fold
+            ld[n] -= 1.0
+            ij = ind[j, ni - 1 - i]
+            if ij >= 0:
+                A[n, ij] = 1.0
+            else:
+                b[n] -= fdata[j, ni - 1 - i]
+    # Set leading diagonal
+    b[ld >= 0] = 0.0
+    A[range(n_missing), range(n_missing)] = ld - stabilizer
+
+    A = scipy.sparse.csr_matrix(A)
+    new_data = np.ma.array(fdata, mask=(mask == 0))
+    if maxiter is None:
+        x, info = scipy.sparse.linalg.bicg(A, b)
+    elif maxiter == 0:
+        x = scipy.sparse.linalg.spsolve(A, b)
     else:
-        raise ValueError(
-            f"{name} values are not consistently in either [-180, 180] or [0, 360] range.\n"
-            f"Min: {lons.min()}, Max: {lons.max()}"
-        )
+        x, info = scipy.sparse.linalg.bicg(A, b, maxiter=maxiter)
+    new_data[missing_j, missing_i] = x
+    return new_data
